Refactor some code
* Use gorilla middleware model for logging
* Add helper function for internal server error response just to remove
  some repetition
